chore: add dynamic pdf viewer wrapper
